Experiment/Agents/Producer/requirements.txt
absl-py==0.11.0
-e git+https://github.com/friedforfun/Dissertation.git@4f281307641ec3d6da3383edb597591efeb619b4#egg=AgentBase&subdirectory=Experiment/CommonFunctionality
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
bqplot==0.11.5
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
configparser==5.0.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
-e git+https://github.com/friedforfun/distiller.git@f401db9403c1608e10258733e013b6c087e3ef34#egg=distiller
docker-pycreds==0.4.0
entrypoints==0.3
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.0
google-pasta==0.2.0
graphviz==0.10.1
grpcio==1.34.0
gym==0.12.5
h5py==3.1.0
idna==2.10
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonpatch==1.28
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
more-itertools==8.6.0
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.18.5
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
pluggy==0.13.1
pretrainedmodels==0.7.4
-e git+https://github.com/friedforfun/Dissertation.git@4f281307641ec3d6da3383edb597591efeb619b4#egg=ProducerAgent&subdirectory=Experiment/Agents/Producer
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
py==1.10.0
pycparser==2.20
pydot==1.4.1
pyglet==1.5.13
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==4.6.11
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.4.1
pyzmq==20.0.0
qgrid==1.1.1
qtconsole==5.0.1
QtPy==1.9.0
redis==3.5.3
requests==2.25.1
scikit-learn==0.21.2
scipy==1.5.4
Send2Trash==1.5.0
sentry-sdk==0.19.5
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
tabulate==0.8.3
tensorboard==1.15.0
tensorflow==1.15.4
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
torch==1.3.1
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.4.2
tornado==6.1
tqdm==4.33.0
traitlets==5.0.5
traittypes==0.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
visdom==0.1.8.9
wandb==0.10.17
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.4.2
wrapt==1.12.1
XlsxWriter==1.3.7
zipp==3.4.0
